fix: notify_bind 从 tools_status 搬到 tools_notify
list_candidates/switch_target/MILOCO_NOTIFY_BIND_SCHEMA/handle_notify_bind
数据归属在 tools_notify(load_state/save_state),操作逻辑应在同一文件
